Buffer rework;
- Instead of flags, Buffer must have 1 type (vertex, index, uniform, storage)
- Transient flag behaves like write flag (temporary per-frame writes)
- Transient buffers are just temporary handles into the scratchpad pool
- Non-transient buffers can be written to now (stages a copy)
- Storage buffers can not be transient
- Parameter flag remains the same, only works with storage buffers
- Change Buffer:hasFlags to Buffer:isTransient
- Add Buffer:drop to reinitialize transient buffer storage
- gpu: u32 buffer offsets/sizes
- gpu: remove gpu_map_buffer, use a scratchpad
- gpu: when grabbing new scratchpad, only check fence when needed
- Call lovr.graphics.begin before lovr.load

+++ src/api/l_graphics_texture.c
@@ -126,7 +126,7 @@ typedef struct {
   uint32_t height;
 } luax_readback;
 
-static void onReadback(void* data, uint64_t size, void* context) {
+static void onReadback(void* data, uint32_t size, void* context) {
   luax_readback* readback = context;
 
   Image* image = lovrImageCreate(readback->width, readback->height, NULL, 0, readback->format);
